Move : MainPage file Rename Home and SignUp page render
  - App.js에 작성했던 메인페이지 코드 Components/Home.js 로 코드이동
  - App.js에 '회원가입' 클릭시 해당 페이지로 이동되도록 코드 추가
