+++ Client/src/Components/Home.css
@@ -1,4 +1,4 @@
-.App {
+.Home {
   text-align: center;
 }
